pyproject.toml
[project]
name = "chatbot-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "faststream[cli,confluent,kafka,otel]>=0.6.6",
    "kafka>=1.3.5",
    "kafka-python>=2.3.0",
    "kafka-python-ng>=2.2.3",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-kafka-python>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "redis>=7.2.0",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]
